backend/requirements.txt
Flask==3.0.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
Flask-Limiter==3.5.*
Flask-CORS==4.0.*
Flask-Mail==0.10.*
SQLAlchemy==2.0.*
psycopg2-binary==2.9.*
alembic==1.13.*
PyJWT==2.8.*
passlib[bcrypt]==1.7.4
bcrypt==4.0.*
marshmallow==3.21.*
python-dotenv==1.0.*
gunicorn==21.2.*
limits==3.7.*
